updated the service class
