+++ Problem-3/src/main/java/com/problem3/business/validation/UrlValidation.java
@@ -1,0 +1,5 @@
+package com.problem3.business.validation;
+
+public class UrlValidation {
+
+}
